[FEATURE] add Command to run the IndexQueue for multiple sites
This feature is a almost direct copy from the same Task in EXT:jwtools2.
The reason for this copy is to make debugging easier by having it as a Command CLI Tool, that uses the Logging Interface to get more and direct informations during the index run. It can be run via the Planer/Scheduler as well.